[Data Liberation] Merge both XML processors into a single WP_XML_Processor
Description TBD
